Exception handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.config.ConfigDataResourceNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
@@ -24,7 +28,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById (Long id) {
-        Optional<Product> result = repository.findById(id);
+        Optional<Product> result = Optional.ofNullable(repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado")));
         Product product = result.get();
         ProductDTO dto = new ProductDTO(product.getId(), product.getName(), product.getDescription(),
                 product.getPrice(), product.getImgUrL());
@@ -83,11 +88,22 @@ public class ProductService {
                 product.getImgUrL()
         );
     }
-    @Transactional
-   public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
-}
+
+        }
+
+
+
